Move test for dereference_node() to test_nodes.py
