+++ tests/test_omegaconf.py
@@ -12,7 +12,6 @@ from omegaconf import (
     IntegerNode,
     ListConfig,
     MissingMandatoryValue,
-    Node,
     OmegaConf,
     StringNode,
 )
@@ -135,23 +134,6 @@ def test_is_missing_resets() -> None:
     assert not OmegaConf.is_missing(cfg, "list")
     cfg.list = "???"
     assert OmegaConf.is_missing(cfg, "list")
-
-
-def test_dereference_missing() -> None:
-    cfg = OmegaConf.create({"x": "???"})
-    x_node = cfg._get_node("x")
-    assert isinstance(x_node, Node)
-    with pytest.raises(MissingMandatoryValue):
-        x_node._dereference_node(throw_on_missing=True)
-
-
-def test_dereference_interpolation_to_missing() -> None:
-    cfg = OmegaConf.create({"x": "${y}", "y": "???"})
-    x_node = cfg._get_node("x")
-    assert isinstance(x_node, Node)
-    assert x_node._dereference_node(throw_on_resolution_failure=False) is None
-    with pytest.raises(MissingMandatoryValue):
-        cfg.x
 
 
 @pytest.mark.parametrize(
